+++ CHANGELOG.md
@@ -1,4 +1,27 @@
 # Change Log
+
+## v2.2.4
+
+### Added
+
+* `forceDragToCreateMode` example in storybook
+* `forceDragToCreateMode` test case
+
+### Changed
+
+* [Modify selection trigger: from `click` to `mouseDown`](https://github.com/flower-platform/react-timeline-10000/pull/50)
+
+### Fixed
+
+* Drag to create rectangle was blue instead of green when `forceDragToCreateMode = true` 
+* [`selectedItems` parameter of `onSelectionChange` handler was `number[]` instead of `(number|string)[]`](https://github.com/flower-platform/react-timeline-10000/compare/master-flower-platform...flower-platform:react-timeline-10000:dana-RM31940-selectedItem-type-number-instead-of-number-or-string?expand=1)
+
+## v2.2.3
+
+### Fixed
+
+Right click selection didn't worked on custom segments with a complex children hierarchy
+
 ## v2.2.2
 
 ### Fixed